+++ src/connectedvmware/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+0.1.9
+++++++
+* Update API Version from 2020-10-01-preview to 2022-01-10-preview.
+* Support for VM delete in retain mode.
+
 0.1.8
 ++++++
 * Displaying asterisks (*****) for password input.